AMP-8712
git-svn-id: file:///svn/branches/AMP_1_14_RELEASE/amp@12803 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/contentrepository/action/DocToOrgAction.java
@@ -97,7 +97,12 @@ public class DocToOrgAction extends MultiAction {
 		
 		DocToOrgForm docToOrgForm			= (DocToOrgForm) form;
 		
-		DocToOrgDAO.deleteDocToOrgObjs(docToOrgForm.getRemovingUuid(), docToOrgForm.getRemovingOrgId() );
+		if(isLoggeedIn(request)){
+			Node n		= DocumentManagerUtil.getReadNode(docToOrgForm.getRemovingUuid(), request);
+			if (n != null && DocumentManagerRights.hasAddParticipatingOrgRights(n, request)) {
+				DocToOrgDAO.deleteDocToOrgObjs(docToOrgForm.getRemovingUuid(), docToOrgForm.getRemovingOrgId() );
+			}
+		}
 		
 		docToOrgForm.setRemovingUuid(null);
 		docToOrgForm.setRemovingOrgId(null);
